Regroup archunit tests around currentContext override
